+++ service/src/main/java/de/citec/dal/hal/unit/PowerConsumptionSensorController.java
@@ -30,7 +30,7 @@ public class PowerConsumptionSensorController extends AbstractUnitController<Pow
         super(config, PowerConsumptionSensorController.class, device, builder);
     }
 
-    public void updatePowerConsumption(final Float value) throws CouldNotPerformException {
+    public void updatePowerConsumption(final Double value) throws CouldNotPerformException {
         logger.debug("Apply power consumption Update[" + value + "] for " + this + ".");
 
         try (ClosableDataBuilder<PowerConsumptionSensor.Builder> dataBuilder = getDataBuilder(this)) {
@@ -41,7 +41,7 @@ public class PowerConsumptionSensorController extends AbstractUnitController<Pow
     }
 
     @Override
-    public float getPowerConsumption() throws NotAvailableException {
+    public Double getPowerConsumption() throws NotAvailableException {
         try {
             return getData().getConsumption();
         } catch (CouldNotPerformException ex) {
